requirements.txt
# Dependencias básicas
pymilvus>=2.3.0
python-dotenv>=1.0.0
openai>=1.0.0

# Procesamiento de documentos
PyMuPDF>=1.23.0
python-docx>=1.0.0

# Utilidades
pydantic>=2.0.0

# Testing
pytest>=7.4.0

